+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 * Add personal data disclosure form (UIDATIMP-786)
 * Adjust "vendor name" field value (UIDATIMP-868)
 * Cover `<FolioRecordTypeSelect>` component with tests (UIDATIMP-705)
+* Cover `<CodeHighlight>` component with tests (UIDATIMP-702)
 
 ### Bugs fixed:
 * Fix Accessibility problems for settings/data-import/match-profiles (lists must only directly contain li elements) (UIDATIMP-452)